+++ src/p99_bluesky/devices/epics/soft_ioc/softsignal.py
@@ -1,3 +1,21 @@
+from softioc import builder
+
+
+def soft_signal(prefix: str, input_name: str, readback_name: str) -> None:
+    # Create some records
+    builder.SetDeviceName(prefix)
+    temp = builder.aIn(readback_name, initial_value=5)
+    # rbv.append(temp)
+    builder.aOut(
+        input_name,
+        initial_value=0.1,
+        always_update=True,
+        on_update=lambda v: temp.set(v),
+    )
+    # builder.device.ai.
+
+
+"""
 import asyncio
 
 from softioc import asyncio_dispatcher, builder, softioc
@@ -31,3 +49,6 @@ dispatcher(update)
 
 # Finally leave the IOC running with an interactive shell.
 softioc.interactive_ioc(globals())
+
+
+"""
